refactor: split RediSearchIndexer and extract AbstractRedisQuery base class
RediSearchIndexer (2365 → 1283 lines) is split into three focused classes:
- SchemaFieldFactory: stateless annotation → Jedis SchemaField converters
- IndexDefinitionBuilder: field routing, nested/reference assembly, delegates to SchemaFieldFactory
- RediSearchIndexer: index lifecycle, keyspace registry, SpEL evaluation

RediSearchQuery (1509 → 963) and RedisEnhancedQuery (1081 → 504) now extend
AbstractRedisQuery (592 lines), which holds all shared fields, the annotation-
processing constructor block (initFromMethod + handleSpecialQueryMethod hook),
and identical execution methods: execute, getQueryMethod, processPartTree,
extractQueryFields, executeAggregation, executeFtTagVals, hasIndexMissing.
~1100 lines of duplication removed across the two query classes.
