+++ redis-om-spring/src/main/java/com/redis/om/spring/indexing/RediSearchIndexer.java
@@ -2,9 +2,6 @@ package com.redis.om.spring.indexing;
 
 import static com.redis.om.spring.util.ObjectUtils.*;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -12,6 +9,7 @@ import java.time.OffsetDateTime;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.BooleanSupplier;
+import java.util.function.Consumer;
 
 import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.logging.Log;
@@ -36,22 +34,16 @@ import org.springframework.util.ClassUtils;
 
 import com.github.f4b6a3.ulid.Ulid;
 import com.google.gson.GsonBuilder;
-import com.google.gson.annotations.JsonAdapter;
-import com.google.gson.annotations.SerializedName;
 import com.redis.om.spring.RedisOMProperties;
 import com.redis.om.spring.annotations.*;
-import com.redis.om.spring.id.IdFilter;
 import com.redis.om.spring.id.IdentifierFilter;
 import com.redis.om.spring.ops.RedisModulesOperations;
 import com.redis.om.spring.ops.search.SearchOperations;
-import com.redis.om.spring.repository.query.QueryUtils;
-import com.redis.om.spring.serialization.gson.EnumTypeAdapter;
 import com.redis.om.spring.tuple.Pair;
 import com.redis.om.spring.tuple.Tuples;
 
 import redis.clients.jedis.exceptions.JedisDataException;
 import redis.clients.jedis.search.FTCreateParams;
-import redis.clients.jedis.search.FieldName;
 import redis.clients.jedis.search.IndexDataType;
 import redis.clients.jedis.search.schemafields.*;
 
@@ -105,6 +97,8 @@ public class RediSearchIndexer {
   private final GsonBuilder gsonBuilder;
   private final RedisOMProperties properties;
   private final ExpressionParser spelParser = new SpelExpressionParser();
+  private final SchemaFieldFactory schemaFieldFactory;
+  private final IndexDefinitionBuilder indexDefinitionBuilder;
 
   /**
    * Constructs a new RediSearchIndexer with the required dependencies.
@@ -122,6 +116,11 @@ public class RediSearchIndexer {
     this.gsonBuilder = gsonBuilder;
     this.rmo = rmo;
     this.mappingContext = mappingContext;
+    this.schemaFieldFactory = new SchemaFieldFactory();
+    this.indexDefinitionBuilder = new IndexDefinitionBuilder(schemaFieldFactory, gsonBuilder,
+        pair -> entityClassToIdentifierFilter.put(pair.getFirst(), pair.getSecond()),
+        (entityClass, fieldName) -> entityClassToLexicographicFields.computeIfAbsent(entityClass,
+            k -> new HashSet<>()).add(fieldName));
   }
 
   /**
@@ -244,10 +243,9 @@ public class RediSearchIndexer {
   // ---------------------------------------------------------------------------
 
   /**
-   * Runs the common field processing pipeline for an entity class: executes
-   * {@link #processIndexedFields}, registers per-field aliases via {@link #registerAlias}
-   * and appends ID-field indexed fields via {@link #createIndexedFieldsForIdFields}. The
-   * returned list contains every {@link SearchField} the caller needs to turn into
+   * Runs the common field processing pipeline for an entity class: delegates to
+   * {@link IndexDefinitionBuilder#buildSearchFields} and registers per-field aliases.
+   * The returned list contains every {@link SearchField} the caller needs to turn into
    * {@link SchemaField}s.
    *
    * @param entityClass    the entity class being indexed
@@ -257,12 +255,10 @@ public class RediSearchIndexer {
    */
   private List<SearchField> prepareSearchFields(Class<?> entityClass, boolean isDocument,
       List<java.lang.reflect.Field> allClassFields) {
-    List<SearchField> searchFields = processIndexedFields(allClassFields, isDocument);
+    List<SearchField> searchFields = indexDefinitionBuilder.buildSearchFields(entityClass, isDocument, allClassFields);
     for (SearchField field : searchFields) {
       registerAlias(entityClass, field.getField().getName(), field.getSchemaField().getFieldName().getAttribute());
     }
-    createIndexedFieldsForIdFields(entityClass, searchFields.stream().map(SearchField::getSchemaField).toList(),
-        isDocument).forEach(searchFields::add);
     return searchFields;
   }
 
@@ -382,7 +378,7 @@ public class RediSearchIndexer {
 
       final List<java.lang.reflect.Field> allClassFields = getDeclaredFieldsTransitively(cl);
       List<SearchField> searchFields = prepareSearchFields(cl, isDocument, allClassFields);
-      Optional<String> maybeScoreField = getDocumentScoreField(allClassFields, isDocument);
+      Optional<String> maybeScoreField = indexDefinitionBuilder.getDocumentScoreField(allClassFields, isDocument);
 
       SearchOperations<String> opsForSearch = rmo.opsForSearch(indexName);
       FTCreateParams params = createIndexDefinition(cl, idxType);
@@ -740,700 +736,6 @@ public class RediSearchIndexer {
     return entityClassToSchema.get(entityClass);
   }
 
-  private List<SearchField> findIndexFields(java.lang.reflect.Field field, String prefix, boolean isDocument) {
-    List<SearchField> fields = new ArrayList<>();
-
-    if (field.isAnnotationPresent(Indexed.class)) {
-      logger.info(String.format("Found @Indexed annotation on field of type: %s", field.getType()));
-
-      Indexed indexed = field.getAnnotation(Indexed.class);
-
-      // Track lexicographic fields
-      if (indexed.lexicographic()) {
-        Class<?> entityClass = field.getDeclaringClass();
-        entityClassToLexicographicFields.computeIfAbsent(entityClass, k -> new HashSet<>()).add(field.getName());
-        logger.info(String.format("Tracked lexicographic field %s on class %s", field.getName(), entityClass
-            .getName()));
-      }
-
-      Class<?> fieldType = ClassUtils.resolvePrimitiveIfNecessary(field.getType());
-
-      if (field.isAnnotationPresent(Reference.class)) {
-        //
-        // @Reference @Indexed fields: Create schema field for the reference entity @Id
-        // field
-        //
-        logger.debug("🪲Found @Reference field " + field.getName() + " in " + field.getDeclaringClass()
-            .getSimpleName());
-        createIndexedFieldForReferenceIdField(field, isDocument).ifPresent(fields::add);
-
-        // Also create index fields for the referenced entity's indexed/searchable fields
-        // This enables searching like RefVehicle$.OWNER_NAME.eq("John")
-        createIndexedFieldsForReferencedEntity(field, isDocument, prefix).forEach(fields::add);
-      } else if (indexed.schemaFieldType() == SchemaFieldType.AUTODETECT) {
-        //
-        // Any Character class, Boolean or Enum with AUTODETECT -> Tag Search Field
-        // Also UUID and Ulid (classes whose toString() is a valid text representation
-        // of the value)
-        //
-        if (CharSequence.class.isAssignableFrom(fieldType) || //
-            (fieldType == Boolean.class) || (fieldType == UUID.class) || (fieldType == Ulid.class)) {
-          fields.add(SearchField.of(field, indexAsTagFieldFor(field, isDocument, prefix, indexed.sortable(), indexed
-              .separator(), indexed.arrayIndex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-        } else if (fieldType.isEnum()) {
-          if (Objects.requireNonNull(indexed.serializationHint()) == SerializationHint.ORDINAL) {
-            fields.add(SearchField.of(field, indexAsNumericFieldFor(field, isDocument, prefix, indexed.sortable(),
-                indexed.noindex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-            gsonBuilder.registerTypeAdapter(fieldType, EnumTypeAdapter.of(fieldType));
-          } else {
-            fields.add(SearchField.of(field, indexAsTagFieldFor(field, isDocument, prefix, indexed.sortable(), indexed
-                .separator(), indexed.arrayIndex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-          }
-        }
-        //
-        // Any Numeric class -> Numeric Search Field
-        //
-        else if ( //
-        Number.class.isAssignableFrom(fieldType) || //
-            (fieldType == LocalDateTime.class) || //
-            (field.getType() == LocalDate.class) || //
-            (field.getType() == Date.class) || //
-            (field.getType() == Instant.class) || //
-            (field.getType() == OffsetDateTime.class) //
-        ) {
-          fields.add(SearchField.of(field, indexAsNumericFieldFor(field, isDocument, prefix, indexed.sortable(), indexed
-              .noindex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-        }
-        //
-        // Set / List
-        //
-        else if (Set.class.isAssignableFrom(fieldType) || List.class.isAssignableFrom(fieldType)) {
-          Optional<Class<?>> maybeCollectionType = getCollectionElementClass(field);
-
-          if (maybeCollectionType.isPresent()) {
-            // https://redis.io/docs/stack/search/indexing_json/#index-limitations
-            // JSON array:
-            // - Array of strings as TAG or TEXT.
-            // - Array of numbers as NUMERIC or VECTOR.
-            // - Array of geo coordinates as GEO.
-            // - null values in such arrays are ignored.
-            Class<?> collectionType = maybeCollectionType.get();
-
-            if (CharSequence.class.isAssignableFrom(collectionType) || (collectionType == Boolean.class)) {
-              fields.add(SearchField.of(field, indexAsTagFieldFor(field, isDocument, prefix, indexed.sortable(), indexed
-                  .separator(), indexed.arrayIndex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-            } else if (isDocument) {
-              if (Number.class.isAssignableFrom(collectionType)) {
-                fields.add(SearchField.of(field, indexAsNumericFieldFor(field, true, prefix, indexed.sortable(), indexed
-                    .noindex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-              } else if (collectionType == Point.class) {
-                fields.add(SearchField.of(field, indexAsGeoFieldFor(field, true, prefix, indexed.alias())));
-              } else if (collectionType == UUID.class || collectionType == Ulid.class) {
-                fields.add(SearchField.of(field, indexAsTagFieldFor(field, true, prefix, indexed.sortable(), indexed
-                    .separator(), 0, indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-              } else {
-                // Index nested JSON fields
-                logger.debug(String.format("Found nested field on field of type: %s", field.getType()));
-                fields.addAll(indexAsNestedFieldFor(field, prefix));
-              }
-            }
-          } else {
-            logger.debug(String.format("Could not determine the type of elements in the collection %s in entity %s",
-                field.getName(), field.getDeclaringClass().getSimpleName()));
-          }
-        }
-        //
-        // Map fields (only for JSON documents)
-        //
-        else if (Map.class.isAssignableFrom(fieldType) && isDocument) {
-          logger.info(String.format("Processing Map field: %s of type %s", field.getName(), fieldType));
-          Optional<Class<?>> maybeValueType = getMapValueClass(field);
-          if (maybeValueType.isPresent()) {
-            Class<?> valueType = maybeValueType.get();
-            logger.info(String.format("Map field %s has value type: %s", field.getName(), valueType));
-
-            // Use the Map field's alias if specified, otherwise use the field name
-            String mapFieldNameForIndex = (indexed.alias() != null && !indexed.alias().isEmpty()) ?
-                indexed.alias() :
-                field.getName();
-
-            String mapJsonPath = (prefix == null || prefix.isBlank()) ?
-                "$." + field.getName() + ".*" :
-                "$." + prefix + "." + field.getName() + ".*";
-            String mapFieldAlias = mapFieldNameForIndex + "_values";
-
-            // Support all value types that we support for regular fields
-            if (CharSequence.class.isAssignableFrom(
-                valueType) || valueType == UUID.class || valueType == Ulid.class || valueType.isEnum()) {
-              // Index as TAG field
-              TagField tagField = TagField.of(FieldName.of(mapJsonPath).as(mapFieldAlias));
-              if (indexed.sortable())
-                tagField.sortable();
-              if (indexed.indexMissing())
-                tagField.indexMissing();
-              if (indexed.indexEmpty())
-                tagField.indexEmpty();
-              if (!indexed.separator().isEmpty()) {
-                tagField.separator(indexed.separator().charAt(0));
-              }
-              fields.add(SearchField.of(field, tagField));
-              logger.info(String.format("Added TAG field for Map: %s as %s", field.getName(), mapFieldAlias));
-            } else if (Number.class.isAssignableFrom(
-                valueType) || valueType == Boolean.class || valueType == LocalDateTime.class || valueType == LocalDate.class || valueType == Date.class || valueType == Instant.class || valueType == OffsetDateTime.class) {
-              // Index as NUMERIC field
-              NumericField numField = NumericField.of(FieldName.of(mapJsonPath).as(mapFieldAlias));
-              if (indexed.sortable())
-                numField.sortable();
-              if (indexed.noindex())
-                numField.noIndex();
-              if (indexed.indexMissing())
-                numField.indexMissing();
-              // NumericField doesn't have indexEmpty() method
-              fields.add(SearchField.of(field, numField));
-              logger.info(String.format("Added NUMERIC field for Map: %s as %s", field.getName(), mapFieldAlias));
-            } else if (valueType == Point.class) {
-              // Index as GEO field
-              GeoField geoField = GeoField.of(FieldName.of(mapJsonPath).as(mapFieldAlias));
-              fields.add(SearchField.of(field, geoField));
-              logger.info(String.format("Added GEO field for Map: %s as %s", field.getName(), mapFieldAlias));
-            } else {
-              // Handle complex object values in Map by recursively indexing their @Indexed fields
-              logger.info(String.format("Processing complex object Map field: %s with value type %s", field.getName(),
-                  valueType.getName()));
-
-              // Recursively process @Indexed fields within the Map value type
-              for (java.lang.reflect.Field subfield : getDeclaredFieldsTransitively(valueType)) {
-                if (subfield.isAnnotationPresent(Indexed.class)) {
-                  Indexed subfieldIndexed = subfield.getAnnotation(Indexed.class);
-                  // Get the actual JSON field name (check for @JsonProperty or @SerializedName)
-                  String jsonFieldName = getJsonFieldName(subfield);
-                  String nestedJsonPath = (prefix == null || prefix.isBlank()) ?
-                      "$." + field.getName() + ".*." + jsonFieldName :
-                      "$." + prefix + "." + field.getName() + ".*." + jsonFieldName;
-                  // Respect the alias annotation on the nested field
-                  String subfieldAlias = (subfieldIndexed.alias() != null && !subfieldIndexed.alias().isEmpty()) ?
-                      subfieldIndexed.alias() :
-                      subfield.getName();
-                  // Use the Map field's alias (if present) for the nested field alias prefix
-                  String nestedFieldAlias = mapFieldNameForIndex + "_" + subfieldAlias;
-
-                  logger.info(String.format("Processing nested field %s in Map value type, path: %s, alias: %s",
-                      subfield.getName(), nestedJsonPath, nestedFieldAlias));
-
-                  Class<?> subfieldType = subfield.getType();
-
-                  // Create appropriate index field based on subfield type
-                  if (CharSequence.class.isAssignableFrom(
-                      subfieldType) || subfieldType == UUID.class || subfieldType == Ulid.class || subfieldType
-                          .isEnum()) {
-                    // Index as TAG field
-                    TagField tagField = TagField.of(FieldName.of(nestedJsonPath).as(nestedFieldAlias));
-                    if (subfieldIndexed.sortable())
-                      tagField.sortable();
-                    if (subfieldIndexed.indexMissing())
-                      tagField.indexMissing();
-                    if (subfieldIndexed.indexEmpty())
-                      tagField.indexEmpty();
-                    if (!subfieldIndexed.separator().isEmpty()) {
-                      tagField.separator(subfieldIndexed.separator().charAt(0));
-                    }
-                    fields.add(SearchField.of(subfield, tagField));
-                    logger.info(String.format("Added nested TAG field for Map value: %s", nestedFieldAlias));
-                  } else if (Number.class.isAssignableFrom(
-                      subfieldType) || subfieldType == Boolean.class || subfieldType == LocalDateTime.class || subfieldType == LocalDate.class || subfieldType == Date.class || subfieldType == Instant.class || subfieldType == OffsetDateTime.class) {
-                    // Index as NUMERIC field
-                    NumericField numField = NumericField.of(FieldName.of(nestedJsonPath).as(nestedFieldAlias));
-                    if (subfieldIndexed.sortable())
-                      numField.sortable();
-                    if (subfieldIndexed.noindex())
-                      numField.noIndex();
-                    if (subfieldIndexed.indexMissing())
-                      numField.indexMissing();
-                    fields.add(SearchField.of(subfield, numField));
-                    logger.info(String.format("Added nested NUMERIC field for Map value: %s", nestedFieldAlias));
-                  } else if (subfieldType == Point.class) {
-                    // Index as GEO field
-                    GeoField geoField = GeoField.of(FieldName.of(nestedJsonPath).as(nestedFieldAlias));
-                    fields.add(SearchField.of(subfield, geoField));
-                    logger.info(String.format("Added nested GEO field for Map value: %s", nestedFieldAlias));
-                  }
-                }
-              }
-            }
-          }
-        }
-        //
-        // Point
-        //
-        else if (fieldType == Point.class) {
-          fields.add(SearchField.of(field, indexAsGeoFieldFor(field, isDocument, prefix, indexed.alias())));
-        }
-        //
-        // Recursively explore the fields for Index annotated fields
-        //
-        else {
-          for (java.lang.reflect.Field subfield : getDeclaredFieldsTransitively(field.getType())) {
-            String subfieldPrefix = (prefix == null || prefix.isBlank()) ?
-                field.getName() :
-                String.join(".", prefix, field.getName());
-            fields.addAll(findIndexFields(subfield, subfieldPrefix, isDocument));
-          }
-        }
-      } else { // Schema field type hardcoded/set in @Indexed
-        switch (indexed.schemaFieldType()) {
-          case TAG -> fields.add(SearchField.of(field, indexAsTagFieldFor(field, isDocument, prefix, indexed.sortable(),
-              indexed.separator(), indexed.arrayIndex(), indexed.alias(), indexed.indexMissing(), indexed
-                  .indexEmpty())));
-          case NUMERIC -> fields.add(SearchField.of(field, indexAsNumericFieldFor(field, isDocument, prefix, indexed
-              .sortable(), indexed.noindex(), indexed.alias(), indexed.indexMissing(), indexed.indexEmpty())));
-          case GEO -> fields.add(SearchField.of(field, indexAsGeoFieldFor(field, true, prefix, indexed.alias())));
-          case VECTOR -> fields.add(SearchField.of(field, indexAsVectorFieldFor(field, isDocument, prefix, indexed)));
-          case NESTED -> {
-            Class<?> nestedType = field.getType();
-
-            // Handle List<Model> fields by extracting the element type
-            if (List.class.isAssignableFrom(nestedType) || Set.class.isAssignableFrom(nestedType)) {
-              Optional<Class<?>> maybeCollectionType = getCollectionElementClass(field);
-              if (maybeCollectionType.isPresent()) {
-                nestedType = maybeCollectionType.get();
-                logger.info(String.format("Processing nested array field %s with element type %s", field.getName(),
-                    nestedType.getSimpleName()));
-              } else {
-                logger.warn(String.format("Could not determine element type for nested field %s", field.getName()));
-                break;
-              }
-            }
-
-            // Process all fields of the nested type automatically
-            for (java.lang.reflect.Field subfield : com.redis.om.spring.util.ObjectUtils.getDeclaredFieldsTransitively(
-                nestedType)) {
-              String subfieldPrefix = (prefix == null || prefix.isBlank()) ?
-                  field.getName() :
-                  String.join(".", prefix, field.getName());
-
-              // For nested fields, automatically create index fields when the parent field is annotated with @Indexed(schemaFieldType = SchemaFieldType.NESTED)
-              fields.addAll(createNestedIndexFields(field, subfield, subfieldPrefix, isDocument));
-            }
-          }
-          default -> {
-          } // NOOP
-        }
-      }
-    }
-
-    // Searchable - behaves like Text indexed
-    else if (field.isAnnotationPresent(Searchable.class)) {
-      logger.info(String.format("Found @Searchable annotation on field of type: %s", field.getType()));
-      Searchable searchable = field.getAnnotation(Searchable.class);
-
-      // Track lexicographic fields
-      if (searchable.lexicographic()) {
-        Class<?> entityClass = field.getDeclaringClass();
-        entityClassToLexicographicFields.computeIfAbsent(entityClass, k -> new HashSet<>()).add(field.getName());
-        logger.info(String.format("Tracked lexicographic field %s on class %s", field.getName(), entityClass
-            .getName()));
-      }
-
-      fields.add(SearchField.of(field, indexAsTextFieldFor(field, isDocument, prefix, searchable)));
-    }
-    // Text
-    else if (field.isAnnotationPresent(TextIndexed.class)) {
-      TextIndexed ti = field.getAnnotation(TextIndexed.class);
-      fields.add(SearchField.of(field, indexAsTextFieldFor(field, isDocument, prefix, ti)));
-    }
-    // Tag
-    else if (field.isAnnotationPresent(TagIndexed.class)) {
-      TagIndexed ti = field.getAnnotation(TagIndexed.class);
-      fields.add(SearchField.of(field, indexAsTagFieldFor(field, isDocument, prefix, ti)));
-    }
-    // Geo
-    else if (field.isAnnotationPresent(GeoIndexed.class)) {
-      GeoIndexed gi = field.getAnnotation(GeoIndexed.class);
-      fields.add(SearchField.of(field, indexAsGeoFieldFor(field, isDocument, prefix, gi)));
-    }
-    // Numeric
-    else if (field.isAnnotationPresent(NumericIndexed.class)) {
-      NumericIndexed ni = field.getAnnotation(NumericIndexed.class);
-      fields.add(SearchField.of(field, indexAsNumericFieldFor(field, isDocument, prefix, ni)));
-    }
-    // Vector
-    else if (field.isAnnotationPresent(VectorIndexed.class)) {
-      VectorIndexed vi = field.getAnnotation(VectorIndexed.class);
-      fields.add(SearchField.of(field, indexAsVectorFieldFor(field, isDocument, prefix, vi)));
-    }
-
-    return fields;
-  }
-
-  private TagField indexAsTagFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix, TagIndexed ti) {
-    FieldName fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(ti.alias()), Optional.empty());
-
-    return getTagField(fieldName, ti.separator(), false);
-  }
-
-  private VectorField indexAsVectorFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      Indexed indexed) {
-    Map<String, Object> attributes = new HashMap<>();
-    attributes.put("TYPE", indexed.type().toString());
-    attributes.put("DIM", indexed.dimension());
-    attributes.put("DISTANCE_METRIC", indexed.distanceMetric());
-
-    if (indexed.initialCapacity() > 0) {
-      attributes.put("INITIAL_CAP", indexed.initialCapacity());
-    }
-
-    // Optional parameters for FLAT
-    if (indexed.algorithm().equals(VectorField.VectorAlgorithm.FLAT) && (indexed.blockSize() > 0)) {
-      attributes.put("BLOCK_SIZE", indexed.blockSize());
-    }
-
-    if (indexed.algorithm().equals(VectorField.VectorAlgorithm.HNSW)) {
-      // Optional parameters for HNSW
-      attributes.put("M", indexed.m());
-      attributes.put("EF_CONSTRUCTION", indexed.efConstruction());
-      if (indexed.efRuntime() != 10) {
-        attributes.put("EF_RUNTIME", indexed.efRuntime());
-      }
-      if (indexed.epsilon() != 0.01) {
-        attributes.put("EPSILON", indexed.epsilon());
-      }
-    }
-
-    FieldName fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(indexed.alias()), Optional
-        .empty());
-
-    return new VectorField(fieldName, indexed.algorithm(), attributes);
-  }
-
-  private VectorField indexAsVectorFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      VectorIndexed vi) {
-    Map<String, Object> attributes = new HashMap<>();
-    attributes.put("TYPE", vi.type().toString());
-    attributes.put("DIM", vi.dimension());
-    attributes.put("DISTANCE_METRIC", vi.distanceMetric());
-
-    if (vi.initialCapacity() > 0) {
-      attributes.put("INITIAL_CAP", vi.initialCapacity());
-    }
-
-    // Optional parameters for FLAT
-    if (vi.algorithm().equals(VectorField.VectorAlgorithm.FLAT) && (vi.blockSize() > 0)) {
-      attributes.put("BLOCK_SIZE", vi.blockSize());
-    }
-
-    if (vi.algorithm().equals(VectorField.VectorAlgorithm.HNSW)) {
-      // Optional parameters for HNSW
-      attributes.put("M", vi.m());
-      attributes.put("EF_CONSTRUCTION", vi.efConstruction());
-      if (vi.efRuntime() != 10) {
-        attributes.put("EF_RUNTIME", vi.efRuntime());
-      }
-      if (vi.epsilon() != 0.01) {
-        attributes.put("EPSILON", vi.epsilon());
-      }
-    }
-
-    FieldName fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(vi.alias()), Optional.empty());
-
-    return new VectorField(fieldName, vi.algorithm(), attributes);
-  }
-
-  private SchemaField indexAsTagFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      boolean sortable, String separator, int arrayIndex, String annotationAlias) {
-    return indexAsTagFieldFor(field, isDocument, prefix, sortable, separator, arrayIndex, annotationAlias, false,
-        false);
-  }
-
-  private SchemaField indexAsTagFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      boolean sortable, String separator, int arrayIndex, String annotationAlias, boolean indexMissing,
-      boolean indexEmpty) {
-    FieldName fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(annotationAlias), Optional.of(
-        arrayIndex));
-    return getTagField(fieldName, separator, sortable, indexMissing, indexEmpty);
-  }
-
-  private TextField indexAsTextFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      TextIndexed ti) {
-    var fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(ti.alias()), Optional.empty());
-    String phonetic = ObjectUtils.isEmpty(ti.phonetic()) ? null : ti.phonetic();
-    return getTextField(fieldName, ti.weight(), ti.sortable(), ti.nostem(), ti.noindex(), phonetic, ti.indexMissing(),
-        ti.indexEmpty());
-  }
-
-  private TextField indexAsTextFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      Searchable ti) {
-    var fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(ti.alias()), Optional.empty());
-    String phonetic = ObjectUtils.isEmpty(ti.phonetic()) ? null : ti.phonetic();
-    return getTextField(fieldName, ti.weight(), ti.sortable(), ti.nostem(), ti.noindex(), phonetic, ti.indexMissing(),
-        ti.indexEmpty());
-  }
-
-  private GeoField indexAsGeoFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix, GeoIndexed gi) {
-    var fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(gi.alias()), Optional.empty());
-    return GeoField.of(fieldName);
-  }
-
-  private NumericField indexAsNumericFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      NumericIndexed ni) {
-    var fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(ni.alias()), Optional.empty());
-    return NumericField.of(fieldName);
-  }
-
-  private NumericField indexAsNumericFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      boolean sortable, boolean noIndex, String annotationAlias) {
-    return indexAsNumericFieldFor(field, isDocument, prefix, sortable, noIndex, annotationAlias, false, false);
-  }
-
-  private NumericField indexAsNumericFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      boolean sortable, boolean noIndex, String annotationAlias, boolean indexMissing, boolean indexEmpty) {
-    var fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(annotationAlias), Optional.empty());
-
-    NumericField num = NumericField.of(fieldName);
-    if (sortable)
-      num.sortable();
-    if (noIndex)
-      num.noIndex();
-    if (indexMissing)
-      num.indexMissing();
-    // Note: NumericField doesn't support indexEmpty() in current Jedis version
-    // if (indexEmpty)
-    //   num.indexEmpty();
-    return num;
-  }
-
-  private GeoField indexAsGeoFieldFor(java.lang.reflect.Field field, boolean isDocument, String prefix,
-      String annotationAlias) {
-    var fieldName = buildFieldName(field, prefix, isDocument, Optional.ofNullable(annotationAlias), Optional.empty());
-    return GeoField.of(fieldName);
-  }
-
-  private List<SearchField> indexAsNestedFieldFor(java.lang.reflect.Field field, String prefix) {
-    String fieldPrefix = getFieldPrefix(prefix, true);
-    return getNestedField(fieldPrefix, field, prefix, null);
-  }
-
-  private List<SearchField> getNestedField(String fieldPrefix, java.lang.reflect.Field field, String prefix,
-      List<SearchField> fieldList) {
-    if (fieldList == null) {
-      fieldList = new ArrayList<>();
-    }
-    Type genericType = field.getGenericType();
-    if (genericType instanceof ParameterizedType pt) {
-      Class<?> actualTypeArgument = (Class<?>) pt.getActualTypeArguments()[0];
-      List<java.lang.reflect.Field> subDeclaredFields = com.redis.om.spring.util.ObjectUtils
-          .getDeclaredFieldsTransitively(actualTypeArgument);
-      String tempPrefix = "";
-      if (prefix == null) {
-        prefix = field.getName();
-      } else {
-        prefix += "." + field.getName();
-      }
-      for (java.lang.reflect.Field subField : subDeclaredFields) {
-
-        Optional<Class<?>> maybeCollectionType = getCollectionElementClass(subField);
-
-        String suffix = (maybeCollectionType.isPresent() && (CharSequence.class.isAssignableFrom(maybeCollectionType
-            .get()) || (maybeCollectionType.get() == Boolean.class))) ? "[*]" : "";
-
-        if (subField.isAnnotationPresent(TagIndexed.class)) {
-          TagIndexed ti = subField.getAnnotation(TagIndexed.class);
-          tempPrefix = field.getName() + "[0:].";
-
-          FieldName fieldName = FieldName.of(fieldPrefix + tempPrefix + subField.getName() + suffix);
-          fieldName = fieldName.as(QueryUtils.searchIndexFieldAliasFor(subField, prefix));
-
-          logger.info(String.format("Creating nested relationships: %s -> %s", field.getName(), subField.getName()));
-          fieldList.add(SearchField.of(field, getTagField(fieldName, ti.separator(), false)));
-          continue;
-        } else if (subField.isAnnotationPresent(Indexed.class)) {
-          boolean subFieldIsTagField = (subField.isAnnotationPresent(Indexed.class) && ( //
-          CharSequence.class.isAssignableFrom(subField.getType()) || //
-              (subField.getType() == Boolean.class) || (subField.getType() == UUID.class) || //
-              ( //
-              maybeCollectionType.isPresent() && //
-                  ( //
-                  CharSequence.class.isAssignableFrom(maybeCollectionType.get()) || //
-                      (maybeCollectionType.get() == Boolean.class) //
-                  ) //
-              ) //
-          ) //
-          );
-          if (subFieldIsTagField) {
-            Indexed indexed = subField.getAnnotation(Indexed.class);
-            tempPrefix = field.getName() + "[0:].";
-
-            FieldName fieldName = FieldName.of(fieldPrefix + tempPrefix + subField.getName() + suffix);
-            String alias = QueryUtils.searchIndexFieldAliasFor(subField, prefix);
-            fieldName = fieldName.as(alias);
-
-            logger.info(String.format("Creating nested relationships: %s -> %s", field.getName(), subField.getName()));
-            fieldList.add(SearchField.of(field, getTagField(fieldName, indexed.separator(), false)));
-            continue;
-          } else if (Number.class.isAssignableFrom(subField.getType()) || (subField
-              .getType() == LocalDateTime.class) || (subField.getType() == LocalDate.class) || (subField
-                  .getType() == Date.class)) {
-
-            FieldName fieldName = FieldName.of(fieldPrefix + tempPrefix + subField.getName() + suffix);
-            String alias = QueryUtils.searchIndexFieldAliasFor(subField, prefix);
-            fieldName = fieldName.as(alias);
-
-            logger.info(String.format("Creating nested relationships: %s -> %s", field.getName(), subField.getName()));
-            fieldList.add(SearchField.of(field, NumericField.of(fieldName)));
-          }
-        } else if (subField.isAnnotationPresent(Searchable.class)) {
-          Searchable searchable = subField.getAnnotation(Searchable.class);
-          tempPrefix = field.getName() + "[0:].";
-
-          FieldName fieldName = FieldName.of(fieldPrefix + tempPrefix + subField.getName() + suffix);
-          String alias = QueryUtils.searchIndexFieldAliasFor(subField, prefix);
-          fieldName = fieldName.as(alias);
-
-          logger.info(String.format("Creating TEXT nested relationships: %s -> %s", field.getName(), subField
-              .getName()));
-
-          String phonetic = ObjectUtils.isEmpty(searchable.phonetic()) ? null : searchable.phonetic();
-
-          fieldList.add(SearchField.of(field, getTextField(fieldName, searchable.weight(), searchable.sortable(),
-              searchable.nostem(), searchable.noindex(), phonetic, searchable.indexMissing(), searchable
-                  .indexEmpty())));
-
-          continue;
-        }
-        if (subField.isAnnotationPresent(Indexed.class)) {
-          getNestedField(fieldPrefix + tempPrefix, subField, prefix, fieldList);
-        }
-      }
-    }
-    return fieldList;
-  }
-
-  /**
-   * Determines the appropriate FieldType for a given Java class.
-   * This utility method centralizes the logic for mapping Java types to Redis field types.
-   */
-  private enum FieldTypeMapper {
-    TAG,
-    NUMERIC,
-    GEO,
-    UNSUPPORTED;
-
-    static FieldTypeMapper getFieldType(Class<?> fieldType) {
-      if (CharSequence.class.isAssignableFrom(
-          fieldType) || fieldType == Boolean.class || fieldType == UUID.class || fieldType == Ulid.class || fieldType
-              .isEnum()) {
-        return TAG;
-      } else if (Number.class.isAssignableFrom(
-          fieldType) || fieldType == LocalDateTime.class || fieldType == LocalDate.class || fieldType == Date.class || fieldType == Instant.class || fieldType == OffsetDateTime.class) {
-        return NUMERIC;
-      } else if (fieldType == Point.class) {
-        return GEO;
-      } else {
-        return UNSUPPORTED;
-      }
-    }
-  }
-
-  /**
-   * Creates index fields for nested array elements automatically.
-   * This method handles automatic indexing of all fields within nested objects
-   * when @Indexed(schemaFieldType = SchemaFieldType.NESTED) is used.
-   */
-  private List<SearchField> createNestedIndexFields(java.lang.reflect.Field arrayField,
-      java.lang.reflect.Field nestedField, String prefix, boolean isDocument) {
-    List<SearchField> fields = new ArrayList<>();
-
-    Class<?> nestedFieldType = ClassUtils.resolvePrimitiveIfNecessary(nestedField.getType());
-
-    // For nested arrays, the path should be: $.arrayField[*].nestedField
-    // The prefix already contains the array field name, so we just need [*].nestedField
-    // JSON documents use a "$" prefix to denote the root of the document, while hash structures do not.
-    // The `isDocument` flag determines whether the entity is stored as a JSON document or a hash structure in Redis.
-    // This affects the field path format: JSON documents require "$." as the prefix, while hash structures do not.
-    String fullFieldPath = isDocument ?
-        "$." + arrayField.getName() + "[*]." + nestedField.getName() :
-        arrayField.getName() + "[*]." + nestedField.getName();
-
-    logger.info(String.format("Creating automatic nested field index: %s -> %s", arrayField.getName(), fullFieldPath));
-
-    // Determine field type and create appropriate index field
-    FieldTypeMapper fieldTypeMapper = FieldTypeMapper.getFieldType(nestedFieldType);
-
-    switch (fieldTypeMapper) {
-      case TAG -> {
-        // Create TAG field for strings, booleans, UUIDs, ULIDs, and enums
-        FieldName fieldName = FieldName.of(fullFieldPath);
-        String alias = QueryUtils.searchIndexFieldAliasFor(nestedField, prefix);
-        if (alias != null && !alias.isEmpty()) {
-          fieldName = fieldName.as(alias);
-        }
-        fields.add(SearchField.of(arrayField, getTagField(fieldName, "|", false)));
-      }
-      case NUMERIC -> {
-        // Create NUMERIC field for numbers and dates
-        FieldName fieldName = FieldName.of(fullFieldPath);
-        String alias = QueryUtils.searchIndexFieldAliasFor(nestedField, prefix);
-        if (alias != null && !alias.isEmpty()) {
-          fieldName = fieldName.as(alias);
-        }
-        fields.add(SearchField.of(arrayField, NumericField.of(fieldName)));
-      }
-      case GEO -> {
-        // Create GEO field for Point objects
-        FieldName fieldName = FieldName.of(fullFieldPath);
-        String alias = QueryUtils.searchIndexFieldAliasFor(nestedField, prefix);
-        if (alias != null && !alias.isEmpty()) {
-          fieldName = fieldName.as(alias);
-        }
-        fields.add(SearchField.of(arrayField, GeoField.of(fieldName)));
-      }
-      case UNSUPPORTED -> logger.debug(String.format("Skipping nested field %s of unsupported type %s", nestedField
-          .getName(), nestedFieldType.getSimpleName()));
-    }
-
-    return fields;
-  }
-
-  private TagField getTagField(FieldName fieldName, String separator, boolean sortable) {
-    return getTagField(fieldName, separator, sortable, false, false);
-  }
-
-  private TagField getTagField(FieldName fieldName, String separator, boolean sortable, boolean indexMissing,
-      boolean indexEmpty) {
-    TagField tag = TagField.of(fieldName);
-    if (separator != null) {
-      if (separator.length() != 1) {
-        throw new IllegalArgumentException("Separator '" + separator + "' is not of length 1.");
-      }
-      tag.separator(separator.charAt(0));
-    }
-    if (sortable)
-      tag.sortable();
-    if (indexMissing)
-      tag.indexMissing();
-    if (indexEmpty)
-      tag.indexEmpty();
-    return tag;
-  }
-
-  private TextField getTextField(FieldName fieldName, double weight, boolean sortable, boolean noStem, boolean noIndex,
-      String phonetic, boolean indexMissing, boolean indexEmpty) {
-    TextField text = TextField.of(fieldName);
-    text.weight(weight);
-    if (sortable)
-      text.sortable();
-    if (noStem)
-      text.noStem();
-    if (noIndex)
-      text.noIndex();
-    if (phonetic != null)
-      text.phonetic(phonetic);
-    if (indexMissing)
-      text.indexMissing();
-    if (indexEmpty)
-      text.indexEmpty();
-    return text;
-  }
-
   private String getEntityPrefix(Class<?> cl) {
     String entityPrefix = cl.getName() + ":";
     if (mappingContext.hasPersistentEntityFor(cl)) {
@@ -1487,335 +789,6 @@ public class RediSearchIndexer {
     } else {
       return Optional.empty();
     }
-  }
-
-  private List<SearchField> processIndexedFields(List<java.lang.reflect.Field> allClassFields, boolean isDocument) {
-    List<SearchField> fields = new ArrayList<>();
-    for (java.lang.reflect.Field field : allClassFields) {
-      fields.addAll(findIndexFields(field, null, isDocument));
-    }
-    return fields;
-  }
-
-  private Optional<String> getDocumentScoreField(List<java.lang.reflect.Field> allClassFields, boolean isDocument) {
-    return allClassFields.stream().filter(field -> field.isAnnotationPresent(DocumentScore.class)).findFirst().map(
-        field -> (isDocument ? "$." : "") + field.getName());
-  }
-
-  private boolean isAnnotationPreset(java.lang.reflect.Field idField, List<SchemaField> fields) {
-    return (!idField.isAnnotationPresent(Indexed.class) && !idField.isAnnotationPresent(Searchable.class) && !idField
-        .isAnnotationPresent(TagIndexed.class) && !idField.isAnnotationPresent(TextIndexed.class) && !idField
-            .isAnnotationPresent(NumericIndexed.class) && (fields.stream().noneMatch(f -> f.getName().equals(idField
-                .getName()))));
-  }
-
-  private List<SearchField> createIndexedFieldsForIdFields(Class<?> cl, List<SchemaField> fields, boolean isDocument) {
-    List<SearchField> results = new ArrayList<>();
-    List<java.lang.reflect.Field> idFields = getIdFieldsForEntityClass(cl);
-    for (java.lang.reflect.Field idField : idFields) {
-      // Only auto-index the @Id if not already indexed by the user (gh-135)
-      if (isAnnotationPreset(idField, fields)) {
-        Class<?> idClass = idField.getType();
-        if (idField.getType().isPrimitive()) {
-          String cls = com.redis.om.spring.util.ObjectUtils.getTargetClassName(idClass.getName());
-          Class<?> primitive = ClassUtils.resolvePrimitiveClassName(cls);
-          if (primitive != null) {
-            idClass = ClassUtils.resolvePrimitiveIfNecessary(primitive);
-          }
-        }
-
-        // TODO: determine if we need to pass the alias
-        if (Number.class.isAssignableFrom(idClass)) {
-          results.add(SearchField.of(idField, indexAsNumericFieldFor(idField, isDocument, "", true, false, null)));
-        } else {
-          results.add(SearchField.of(idField, indexAsTagFieldFor(idField, isDocument, "", false, "|", Integer.MIN_VALUE,
-              null)));
-        }
-      }
-    }
-
-    java.lang.reflect.Field idField = (!idFields.isEmpty()) ? idFields.get(0) : null;
-
-    // register any @IdFilter annotation
-    // TODO If multiple IDs which one does the IdFilter applies to? - for now the first
-    if (idField != null && idField.isAnnotationPresent(IdFilter.class)) {
-      IdFilter idFilter = idField.getAnnotation(IdFilter.class);
-      var identifierFilterClass = idFilter.value();
-      try {
-        var identifierFilter = (IdentifierFilter<?>) identifierFilterClass.getDeclaredConstructor().newInstance();
-        entityClassToIdentifierFilter.put(cl, identifierFilter);
-      } catch (InstantiationException | IllegalAccessException | InvocationTargetException |
-               NoSuchMethodException idFilterInstantiationException) {
-        logger.error(String.format("Could not instantiate IdFilter of type %s applied to class %s",
-            identifierFilterClass.getSimpleName(), cl), idFilterInstantiationException);
-      }
-    }
-
-    return results;
-  }
-
-  private Optional<SearchField> createIndexedFieldForReferenceIdField( //
-      java.lang.reflect.Field referenceIdField, //
-      boolean isDocument) {
-    SerializedName serializedName = referenceIdField.getAnnotation(SerializedName.class);
-    String fname = (serializedName != null) ? serializedName.value() : referenceIdField.getName();
-
-    String fieldPrefix = getFieldPrefix("", isDocument);
-    FieldName fieldName = FieldName.of(fieldPrefix + fname);
-    String alias = QueryUtils.searchIndexFieldAliasFor(referenceIdField, "");
-    fieldName = fieldName.as(alias);
-
-    return Optional.of(SearchField.of(referenceIdField, isDocument ?
-        TagField.of(fieldName).separator('|') :
-        TagField.of(fieldName).separator('|').sortable()));
-  }
-
-  /**
-   * Creates index fields for the indexed/searchable fields of a referenced entity.
-   * This enables searching on referenced entity properties, e.g., RefVehicle$.OWNER_NAME.eq("John").
-   *
-   * @param referenceField the @Reference field
-   * @param isDocument     whether this is a JSON document (vs Hash)
-   * @param prefix         the field prefix
-   * @return list of search fields for the referenced entity's indexed properties
-   */
-  private List<SearchField> createIndexedFieldsForReferencedEntity(java.lang.reflect.Field referenceField,
-      boolean isDocument, String prefix) {
-
-    List<SearchField> fields = new ArrayList<>();
-    Class<?> referencedType = referenceField.getType();
-    String referenceFieldName = referenceField.getName();
-
-    logger.debug(
-        "Processing indexed subfields for @Reference field " + referenceFieldName + " of type " + referencedType
-            .getSimpleName());
-
-    // Get all fields from the referenced entity that have indexing annotations
-    List<java.lang.reflect.Field> referencedFields = new ArrayList<>();
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        Indexed.class));
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        Searchable.class));
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        TagIndexed.class));
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        TextIndexed.class));
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        NumericIndexed.class));
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        GeoIndexed.class));
-    referencedFields.addAll(com.redis.om.spring.util.ObjectUtils.getFieldsWithAnnotation(referencedType,
-        VectorIndexed.class));
-    // Remove duplicates (a field might have multiple annotations)
-    referencedFields = referencedFields.stream().distinct().toList();
-
-    for (java.lang.reflect.Field subField : referencedFields) {
-      // Skip @Id fields - they're handled separately by createIndexedFieldForReferenceIdField
-      if (subField.isAnnotationPresent(Id.class)) {
-        continue;
-      }
-      // Skip @Reference fields to avoid infinite recursion
-      if (subField.isAnnotationPresent(Reference.class)) {
-        continue;
-      }
-
-      Class<?> subFieldType = ClassUtils.resolvePrimitiveIfNecessary(subField.getType());
-      String subFieldName = subField.getName();
-
-      // Build the nested field path: referenceField.subField
-      String fieldPath = isDocument ?
-          getFieldPrefix(prefix, true) + referenceFieldName + "." + subFieldName :
-          referenceFieldName + "_" + subFieldName;
-
-      // Build the alias: referenceField_subField
-      String alias = referenceFieldName + "_" + subFieldName;
-
-      FieldName fieldName = FieldName.of(fieldPath).as(alias);
-
-      logger.debug(
-          "Creating index field for " + referenceFieldName + "." + subFieldName + " with path " + fieldPath + " and alias " + alias);
-
-      // Handle @Searchable fields (full-text search)
-      Searchable searchable = subField.getAnnotation(Searchable.class);
-      if (searchable != null) {
-        TextField textField = TextField.of(fieldName);
-        if (searchable.weight() != 1.0) {
-          textField.weight(searchable.weight());
-        }
-        if (searchable.sortable()) {
-          textField.sortable();
-        }
-        if (searchable.nostem()) {
-          textField.noStem();
-        }
-        if (searchable.noindex()) {
-          textField.noIndex();
-        }
-        String phonetic = searchable.phonetic();
-        if (phonetic != null && !phonetic.isEmpty()) {
-          textField.phonetic(phonetic);
-        }
-        if (searchable.indexMissing()) {
-          textField.indexMissing();
-        }
-        if (searchable.indexEmpty()) {
-          textField.indexEmpty();
-        }
-        fields.add(SearchField.of(subField, textField));
-        continue;
-      }
-
-      // Handle @TextIndexed fields
-      TextIndexed textIndexed = subField.getAnnotation(TextIndexed.class);
-      if (textIndexed != null) {
-        TextField textField = TextField.of(fieldName);
-        if (textIndexed.weight() != 1.0) {
-          textField.weight(textIndexed.weight());
-        }
-        if (textIndexed.sortable()) {
-          textField.sortable();
-        }
-        if (textIndexed.nostem()) {
-          textField.noStem();
-        }
-        if (textIndexed.noindex()) {
-          textField.noIndex();
-        }
-        String phonetic = textIndexed.phonetic();
-        if (phonetic != null && !phonetic.isEmpty()) {
-          textField.phonetic(phonetic);
-        }
-        if (textIndexed.indexMissing()) {
-          textField.indexMissing();
-        }
-        if (textIndexed.indexEmpty()) {
-          textField.indexEmpty();
-        }
-        fields.add(SearchField.of(subField, textField));
-        continue;
-      }
-
-      // Handle @Indexed or @TagIndexed fields
-      Indexed indexed = subField.getAnnotation(Indexed.class);
-      TagIndexed tagIndexed = subField.getAnnotation(TagIndexed.class);
-      NumericIndexed numericIndexed = subField.getAnnotation(NumericIndexed.class);
-      GeoIndexed geoIndexed = subField.getAnnotation(GeoIndexed.class);
-
-      if (tagIndexed != null || (indexed != null && (CharSequence.class.isAssignableFrom(
-          subFieldType) || subFieldType == java.util.UUID.class || subFieldType == com.github.f4b6a3.ulid.Ulid.class))) {
-        // Tag field for strings, UUID, and Ulid
-        String separatorStr = tagIndexed != null ?
-            tagIndexed.separator() :
-            (indexed != null ? indexed.separator() : "|");
-        char separator = separatorStr != null && !separatorStr.isEmpty() ? separatorStr.charAt(0) : '|';
-        TagField tagField = TagField.of(fieldName).separator(separator);
-        if (indexed != null && indexed.sortable()) {
-          tagField.sortable();
-        }
-        if (tagIndexed != null && tagIndexed.indexMissing()) {
-          tagField.indexMissing();
-        } else if (indexed != null && indexed.indexMissing()) {
-          tagField.indexMissing();
-        }
-        if (tagIndexed != null && tagIndexed.indexEmpty()) {
-          tagField.indexEmpty();
-        } else if (indexed != null && indexed.indexEmpty()) {
-          tagField.indexEmpty();
-        }
-        fields.add(SearchField.of(subField, tagField));
-      } else if (numericIndexed != null || (indexed != null && (Number.class.isAssignableFrom(
-          subFieldType) || subFieldType == java.time.LocalDateTime.class || subFieldType == java.time.LocalDate.class || subFieldType == java.util.Date.class || subFieldType == java.time.Instant.class || subFieldType == java.time.OffsetDateTime.class))) {
-        // Numeric field
-        NumericField numField = NumericField.of(fieldName);
-        if ((numericIndexed != null && numericIndexed.sortable()) || (indexed != null && indexed.sortable())) {
-          numField.sortable();
-        }
-        if ((numericIndexed != null && numericIndexed.noindex()) || (indexed != null && indexed.noindex())) {
-          numField.noIndex();
-        }
-        if (indexed != null && indexed.indexMissing()) {
-          numField.indexMissing();
-        }
-        // Note: NumericField doesn't support indexEmpty() in current Jedis version
-        fields.add(SearchField.of(subField, numField));
-      } else if (geoIndexed != null || (indexed != null && Point.class.isAssignableFrom(subFieldType))) {
-        // Geo field
-        GeoField geoField = GeoField.of(fieldName);
-        fields.add(SearchField.of(subField, geoField));
-      } else if (indexed != null && subFieldType.isEnum()) {
-        // Enum as tag field
-        String separatorStr = indexed.separator();
-        char separator = separatorStr != null && !separatorStr.isEmpty() ? separatorStr.charAt(0) : '|';
-        TagField tagField = TagField.of(fieldName).separator(separator);
-        if (indexed.sortable()) {
-          tagField.sortable();
-        }
-        if (indexed.indexMissing()) {
-          tagField.indexMissing();
-        }
-        if (indexed.indexEmpty()) {
-          tagField.indexEmpty();
-        }
-        fields.add(SearchField.of(subField, tagField));
-      } else if (indexed != null && (subFieldType == Boolean.class || subFieldType == boolean.class)) {
-        // Boolean as tag field
-        TagField tagField = TagField.of(fieldName);
-        if (indexed.sortable()) {
-          tagField.sortable();
-        }
-        if (indexed.indexMissing()) {
-          tagField.indexMissing();
-        }
-        if (indexed.indexEmpty()) {
-          tagField.indexEmpty();
-        }
-        fields.add(SearchField.of(subField, tagField));
-      }
-
-      // Handle @VectorIndexed fields
-      VectorIndexed vectorIndexed = subField.getAnnotation(VectorIndexed.class);
-      if (vectorIndexed != null) {
-        VectorField.VectorAlgorithm algorithm = vectorIndexed.algorithm();
-        VectorType vectorType = vectorIndexed.type();
-        int dimension = vectorIndexed.dimension();
-        DistanceMetric distanceMetric = vectorIndexed.distanceMetric();
-        int initialCap = vectorIndexed.initialCapacity();
-
-        Map<String, Object> vectorAttrs = new HashMap<>();
-        vectorAttrs.put("TYPE", vectorType.toString());
-        vectorAttrs.put("DIM", dimension);
-        vectorAttrs.put("DISTANCE_METRIC", distanceMetric.toString());
-        if (initialCap > 0) {
-          vectorAttrs.put("INITIAL_CAP", initialCap);
-        }
-
-        if (algorithm == VectorField.VectorAlgorithm.HNSW) {
-          int m = vectorIndexed.m();
-          int efConstruction = vectorIndexed.efConstruction();
-          int efRuntime = vectorIndexed.efRuntime();
-          double epsilon = vectorIndexed.epsilon();
-          if (m > 0)
-            vectorAttrs.put("M", m);
-          if (efConstruction > 0)
-            vectorAttrs.put("EF_CONSTRUCTION", efConstruction);
-          if (efRuntime > 0)
-            vectorAttrs.put("EF_RUNTIME", efRuntime);
-          if (epsilon > 0)
-            vectorAttrs.put("EPSILON", epsilon);
-        } else if (algorithm == VectorField.VectorAlgorithm.FLAT) {
-          int blockSize = vectorIndexed.blockSize();
-          if (blockSize > 0)
-            vectorAttrs.put("BLOCK_SIZE", blockSize);
-        }
-
-        VectorField vectorField = VectorField.builder().fieldName(fieldName).algorithm(algorithm).attributes(
-            vectorAttrs).build();
-
-        fields.add(SearchField.of(subField, vectorField));
-      }
-    }
-
-    return fields;
   }
 
   private FTCreateParams createIndexDefinition(Class<?> cl, IndexDataType idxType) {
@@ -1881,63 +854,8 @@ public class RediSearchIndexer {
     return keyspace.endsWith(":") ? keyspace : keyspace + ":";
   }
 
-  private String getFieldPrefix(String prefix, boolean isDocument) {
-    String chain = (prefix == null || prefix.isBlank()) ? "" : prefix + ".";
-    return isDocument ? "$." + chain : chain;
-  }
-
-  private String getJsonFieldName(java.lang.reflect.Field field) {
-    // Check for @JsonProperty annotation first
-    if (field.isAnnotationPresent(com.fasterxml.jackson.annotation.JsonProperty.class)) {
-      com.fasterxml.jackson.annotation.JsonProperty jsonProperty = field.getAnnotation(
-          com.fasterxml.jackson.annotation.JsonProperty.class);
-      if (jsonProperty.value() != null && !jsonProperty.value().isEmpty()) {
-        return jsonProperty.value();
-      }
-    }
-
-    // Check for @SerializedName annotation (Gson)
-    if (field.isAnnotationPresent(com.google.gson.annotations.SerializedName.class)) {
-      com.google.gson.annotations.SerializedName serializedName = field.getAnnotation(
-          com.google.gson.annotations.SerializedName.class);
-      if (serializedName.value() != null && !serializedName.value().isEmpty()) {
-        return serializedName.value();
-      }
-    }
-
-    // Default to field name
-    return field.getName();
-  }
-
   private void registerAlias(Class<?> cl, String fieldName, String alias) {
     entityClassFieldToAlias.put(Tuples.of(cl, fieldName), alias);
-  }
-
-  private FieldName buildFieldName( //
-      java.lang.reflect.Field field, String prefix, boolean isDocument, Optional<String> maybeAlias,
-      Optional<Integer> maybeArrayIndex) {
-    SerializedName serializedName = field.getAnnotation(SerializedName.class);
-    Indexed indexed = field.getAnnotation(Indexed.class);
-    String fname = (serializedName != null) ? serializedName.value() : field.getName();
-
-    TypeInformation<?> typeInfo = TypeInformation.of(field.getType());
-    String fieldPrefix = getFieldPrefix(prefix, isDocument);
-
-    String index = maybeArrayIndex.isPresent() && (maybeArrayIndex.get() != Integer.MIN_VALUE) ?
-        ".[" + maybeArrayIndex.get() + "]" :
-        "[*]";
-
-    boolean needsPostfix = (isDocument && typeInfo.isCollectionLike() && !field.isAnnotationPresent(
-        JsonAdapter.class) && (indexed != null && !indexed.schemaFieldType().equals(SchemaFieldType.VECTOR)));
-    String fieldPostfix = needsPostfix ? index : "";
-
-    String name = fieldPrefix + fname + fieldPostfix;
-
-    String alias = maybeAlias.isEmpty() || maybeAlias.get().isBlank() ?
-        QueryUtils.searchIndexFieldAliasFor(field, prefix) :
-        maybeAlias.get();
-
-    return FieldName.of(name).as(alias);
   }
 
   /**
@@ -2100,7 +1018,7 @@ public class RediSearchIndexer {
 
       final List<java.lang.reflect.Field> allClassFields = getDeclaredFieldsTransitively(entityClass);
       List<SearchField> searchFields = prepareSearchFields(entityClass, isDocument, allClassFields);
-      Optional<String> maybeScoreField = getDocumentScoreField(allClassFields, isDocument);
+      Optional<String> maybeScoreField = indexDefinitionBuilder.getDocumentScoreField(allClassFields, isDocument);
 
       SearchOperations<String> opsForSearch = rmo.opsForSearch(indexName);
       FTCreateParams params = createIndexDefinition(entityClass, idxType);
@@ -2168,7 +1086,7 @@ public class RediSearchIndexer {
 
       final List<java.lang.reflect.Field> allClassFields = getDeclaredFieldsTransitively(entityClass);
       List<SearchField> searchFields = prepareSearchFields(entityClass, isDocument, allClassFields);
-      Optional<String> maybeScoreField = getDocumentScoreField(allClassFields, isDocument);
+      Optional<String> maybeScoreField = indexDefinitionBuilder.getDocumentScoreField(allClassFields, isDocument);
 
       SearchOperations<String> opsForSearch = rmo.opsForSearch(indexName);
       FTCreateParams params = createIndexDefinition(entityClass, idxType, repoIndexingOptions);
